Remove User, Add popFilter and Buffer
