--- user.go
@@ -1,5 +1,0 @@
-package scry
-
-type User struct {
-	ID string
-}
